test: Sketch out the initial MessengerDriver and TestDriver components.
For now these are super-simple with only a few orders, and
a bunch of those are marked "LATER". The first implementation we're
going to do won't even rely on the Messenger reporting any state
back; it'll just use the existing public interface.
Once that's done and works, we'll use StateTracker in the Messenger
and use that for the stuff marked LATER, then move on to more
sophisticated testing.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>