test: convert StateAlerts to be allocated by TestDriver.
Give them a reference to their lock as well, and let users either
specify a lock and Cond, or generate the data members. This
will make life easier on simple tests (which don't need to handle so
many objects), and on complicated tests (which will want to watch
for multiple alerts at once).

Convert current users.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>